requirements.txt
selenium==4.21.0
webdriver-manager==3.8.6
beautifulsoup4==4.12.2
PyYAML==6.0
pandas==2.1.1
lxml
psutil